+++ core/triggers/interface_90_modSubtotal_subtotaltrigger.class.php
@@ -258,7 +258,7 @@ class Interfacesubtotaltrigger extends DolibarrTriggers
             }
 		    else
             {
-			    $subtotal_add_title_bloc_from_orderstoinvoice = GETPOST('subtotal_add_title_bloc_from_orderstoinvoice', 'none');
+			    $subtotal_add_title_bloc_from_orderstoinvoice = (GETPOST('subtotal_add_title_bloc_from_orderstoinvoice', 'none') || GETPOST('createbills_onebythird', 'int'));
 			    if (!empty($subtotal_add_title_bloc_from_orderstoinvoice))
 			    {
 				    global $subtotal_current_rang, $subtotal_bloc_previous_fk_commande, $subtotal_bloc_already_add_title, $subtotal_bloc_already_add_st;
@@ -405,7 +405,8 @@ class Interfacesubtotaltrigger extends DolibarrTriggers
 			$object->fetchObjectLinked();
 
 			foreach ($object->lines as &$line) {
-
+				$orderline = new OrderLine($this->db);
+				$orderline->fetch($line->origin_line_id);
 				// si la conf pas d'affichage des titres  et consorts (sous total )
 				//on supprime la ligne de sous total
 				if ($conf->global->NO_TITLE_SHOW_ON_EXPED_GENERATION){
@@ -466,7 +467,6 @@ class Interfacesubtotaltrigger extends DolibarrTriggers
 					}
 				}
 			}
-
 			if (!empty($TLinesToDelete)) {
 				foreach ($TLinesToDelete as $lineToDelete) {
 					$lineToDelete->delete($user);